refactor(moqt): share the in-process DUAL session helpers between tests
The DUAL client tests generated a self-signed cert and spun up a server
inline. Domain-level tests need the same setup, so it moves to
modules/test_support.rs.
