+++ moqt/src/modules.rs
@@ -1,4 +1,6 @@
 pub(crate) mod extensions;
 pub(crate) mod moqt;
+#[cfg(all(test, not(target_arch = "wasm32")))]
+pub(crate) mod test_support;
 #[cfg(not(target_arch = "wasm32"))]
 pub(crate) mod transport;
